Don't focusstack when fullscreen
When a client is fullscreen on a monitor, vanilla dwm lets you focus a
different window in the stack. However, the new client does not become
visible because the old client is still fullscreen. This is a problem
because it is possible to make the new "hidden" client fullscreen and
not realize it. Then, when the old client is un-fullscreen-ed, the new
client takes over the entire view and I lose my mind trying to figure
out if I accidentally changed to monocle mode or something.

This commit makes it so that nothing else in the stack can be focused
when the selected client is fullscreen. It is still possible to focus
windows on a different monitor though.

There is currently one side-effect though - when a fullscreen window is
moved to a different monitor, it is moved logically but the view is not
refreshed. The current workaround for this is to un-fullscreen and
re-fullscreen the client. But we probably need to find a better fix for
this.