Apply resetlayout patch
This patch is supposed to reset the layout and mfact when there is only
one client visible. However it does not quite work the way I want it to;
when there is only one client visible the layout seems to get updated
but the screen does not get redrawn. Perhaps it is not calling the
appropriate redraw functions?